refactor(probe): complete dhcp migration with services and cli hooks
